Fixed handling of null values in LocalDateConverter.  By default, org.apache.commons.beanutils.converters.AbstractConverter will throw a ConversionException if the input value is null, unless you've specified a default value.  However, if you specify a default value, it also returns that value instead of throwing any ConversionExceptions which may happen for other reasons.  But that isn't what we want.  Instead, we want the output value to be null when the input value is null, but we still want to get ConversionExceptions when they happen.
Upong looking at the source for AbstractConverter, it seemed that the appropriate thing to do is override handleMissing(), which is called whenever a null input is given, but _before_ convertToType() is invoked.  This gives a chance to decide what to do about null input values, while still allowing ConversionExceptions to happen elsewhere.